chore: extract to_interactive_clap_context_scope_trait module
